refactor(today): extract provenance helpers
